app/requirements.txt
flask
matplotlib
Pillow
requests
